Move `DefaultValueExpression` to expressions
