+++ java/src/main/kotlin/io/spine/validation/java/expression/DefaultValueExpression.kt
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.validation.java
+package io.spine.validation.java.expression
 
 import io.spine.protodata.ast.Field
 import io.spine.protodata.java.Expression
